Allows attackees to fight back and cleans up a bit
