+++ src/adv.py
@@ -4,16 +4,6 @@ from data.characters import pc
 from helpers.general import format_list, names, wrap, flatten_object
 from adventure.actions import handle_get, handle_drop, handle_attack, handle_inspect
 
-# Write a loop that:
-#
-# * Prints the current room name
-# * Prints the current description (the textwrap module might be useful here).
-# * Waits for user input and decides what to do.
-#
-# If the user enters a cardinal direction, attempt to move to the room there.
-# Print an error message if the movement isn't allowed.
-#
-# If the user enters 'q', quit the game.
 while True:
 	print(f'\n{pc.room}')
 	print(wrap(pc.room.description))
