+++ src/helpers/general.py
@@ -13,8 +13,8 @@ def format_list_with_end(l, end = 'or'):
   return f"{format_list(l[:-1])}{',' if len(l) > 2 else ''} {end} {last}"
 
 from random import randint
-def randBool():
-  return randint(0, 1) == 0
+def randBool(n = 1):
+  return randint(0, n) == 0
 
 def flatten_object(o):
   return [item for k, sl in o.items() for item in sl]
